♻️ layout.tsxで共通化する
